python/concert_crawler/requirements.txt
requests==2.32.3
python-dotenv==1.1.0
Pillow==11.2.1
beautifulsoup4==4.13.4
selenium==4.31.0
urllib3==2.4.0
webdriver-manager==4.0.2
boto3==1.38.11
openai==1.77.0